requirements.txt
Django==5.2.7
djangorestframework==3.16.1
djoser==2.3.3
mysqlclient==2.2.7
python-decouple==3.8
django-cors-headers==4.9.0
Pillow==12.0.0
djangorestframework-simplejwt==5.5.1drf-nested-routers==0.95.0
